Removed sacred from schnetpack because limitations sucked
