+++ src/sacred_scripts/__init__.py
@@ -1,4 +1,2 @@
-from sacred_scripts.spk_train import *
-from sacred_scripts.spk_eval import *
 from sacred_scripts.spk_md import *
 from sacred_scripts.spk_parse import *
