+++ src/sacred_scripts/spk_parse.py
@@ -1,9 +1,9 @@
 import os
 from sacred import Experiment
 from schnetpack.data import generate_db
-from schnetpack.sacred.folder_ingredient import save_config, create_dirs, folder_ing
+from shutil import rmtree
 
-parsing = Experiment("parsing", ingredients=[folder_ing])
+parsing = Experiment("parsing")
 
 
 @parsing.config
@@ -17,6 +17,7 @@ def config():
         "Properties=species:S:1:pos:R:3"
     )  # atomic properties of the input file
     molecular_properties = ["energy"]  # molecular properties of the input file
+    overwrite = False
 
 
 @parsing.named_config
@@ -25,6 +26,32 @@ def forces():
     Adds forces to the atomic property string.
     """
     atomic_properties = "Properties=species:S:1:pos:R:3:forces:R:3"
+
+
+@parsing.capture
+def create_dirs(_log, output_dir, overwrite):
+    """
+    Create the directory for the experiment.
+
+    Args:
+        output_dir (str): path to the output directory
+        overwrite (bool): overwrites the model directory if True
+    """
+    _log.info("Create model directory")
+
+    if output_dir is None:
+        raise ValueError("Config `output_dir` has to be set!")
+
+    if os.path.exists(output_dir) and not overwrite:
+        raise ValueError(
+            "Output directory already exists (set overwrite flag?):", output_dir
+        )
+
+    if os.path.exists(output_dir) and overwrite:
+        rmtree(output_dir)
+
+    if not os.path.exists(output_dir):
+        os.makedirs(output_dir)
 
 
 @parsing.command
@@ -41,7 +68,6 @@ def parse(_log, _config, file_path, db_path, atomic_properties, molecular_proper
     """
     output_dir = os.path.dirname(db_path)
     create_dirs(_log=_log, output_dir=output_dir)
-    save_config(_config=_config, output_dir=output_dir)
     generate_db(file_path, db_path, atomic_properties, molecular_properties)
