--- src/schnetpack/sacred/__init__.py
@@ -1,7 +1,0 @@
-"""
-
-"""
-
-from schnetpack.sacred.trainer_ingredients import *
-from schnetpack.sacred.model_ingredients import *
-from schnetpack.sacred.dataset_ingredients import *
